Restructure frontend components and improve file organization
- Moved authentication and contribution form components to more organized directory structure
- Refactored shared components like MyInput, MyCheckbox, and InputError
- Updated import paths across frontend files to reflect new component locations
- Improved type safety and component reusability
- Cleaned up test files to match new component paths
