+++ app/frontend/components/pages/auth/connection/ConnectionForm.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from 'react'
 import useFormHandler from '@/hooks/useFormHandler'
 import { z } from 'zod'
-import MyInput from './MyInput'
+import MyInput from '../../../shared/MyInput'
 import { Button } from '@/components/ui/button'
 import { StepForward } from 'lucide-react'
