+++ app/frontend/components/pages/auth/signIn/SignInForm.tsx
@@ -9,7 +9,7 @@ import {
 } from '@/components/ui/input-otp'
 import { REGEXP_ONLY_DIGITS } from 'input-otp'
 import { z } from 'zod'
-import InputError from '@/components/forms/InputError'
+import InputError from '@/components/shared/InputError'
 import { Button } from '@/components/ui/button'
 import { StepForward } from 'lucide-react'
 import ResendCode from '@/components/pages/auth/signIn/ResendCode'
