+++ app/frontend/components/pages/auth/signUp/SignUpForm.tsx
@@ -2,8 +2,8 @@ import { useRef, ReactElement, FormEvent } from 'react'
 import { Link } from '@inertiajs/react'
 import useFormHandler from '@/hooks/useFormHandler'
 import { KeysWithStringValues } from '@/types/utilityTypes'
-import MyInput from './MyInput'
-import MyCheckbox from './MyCheckbox'
+import MyInput from '../shared/MyInput'
+import MyCheckbox from '../shared/MyCheckbox'
 import ReCAPTCHA from 'react-google-recaptcha'
 import { Button } from '@/components/ui/button'
 import { StepForward } from 'lucide-react'
