+++ app/frontend/components/pages/contribution/new/OsblDocuments/OsblDocumentSheet.tsx
@@ -1,7 +1,7 @@
 import { ReactElement, useRef, useState } from 'react'
 import { Document, FormProps } from '@/pages/Contribution/types'
 import { Button } from '@/components/ui/button'
-import MyInput from '@/components/forms/MyInput'
+import MyInput from '@/components/shared/MyInput'
 import {
   Select,
   SelectContent,
@@ -9,7 +9,7 @@ import {
   SelectTrigger,
   SelectValue
 } from '@/components/ui/select'
-import MyNumberInput from '@/components/forms/MyNumberInput'
+import MyNumberInput from '@/components/shared/MyNumberInput'
 import { Textarea } from '@/components/ui/textarea'
 import MyFileInput from '@/components/shared/MyFileInput'
 import { CheckIcon, ChevronDown, X } from 'lucide-react'
