+++ app/frontend/components/pages/contribution/new/OsblFinances/OsblFinanceSheet.tsx
@@ -1,9 +1,9 @@
 import React, { ReactElement, useRef, useState } from 'react'
 import { AnnualFinance, FormProps } from '@/pages/Contribution/types'
-import MyNumberInput from '@/components/forms/MyNumberInput'
+import MyNumberInput from '@/components/shared/MyNumberInput'
 import HelpTooltip from '@/components/shared/HelpTooltip'
-import FundManagementSection from './FundManagementSection'
-import MyCheckbox from '@/components/forms/MyCheckbox'
+import FundManagementSection from '@/components/pages/contribution/new/OsblFinances/FundManagementSection'
+import MyCheckbox from '@/components/shared/MyCheckbox'
 import { Separator } from '@/components/ui/separator'
 import { Button } from '@/components/ui/button'
 import { CheckIcon, X } from 'lucide-react'
