+++ app/frontend/pages/Auth/Connection.tsx
@@ -1,6 +1,6 @@
 import { Head } from '@inertiajs/react'
 import { ReactElement } from 'react'
-import ConnectionForm from '@/components/forms/ConnectionForm'
+import ConnectionForm from '@/components/pages/auth/connection/ConnectionForm'
 import QuoteSection from '@/components/pages/QuoteSection'
 // @ts-expect-error
 import Superwoman from '@/assets/images/auth/superwoman.svg?react'
