+++ app/frontend/pages/Contribution/New.tsx
@@ -6,8 +6,8 @@ import GoodIdea from '@/assets/icons/good-idea.svg?react'
 import { Button } from '@/components/ui/button'
 import { Save } from 'lucide-react'
 import OsblHeader from '@/components/pages/contribution/new/OsblHeader'
-import OsblDataSheet from '@/components/pages/contribution/new/OsblDataSheet'
-import OsblFinance from '@/components/pages/contribution/new/OsblFinance'
+import OsblDataSheet from '@/components/pages/contribution/new/OsblDatasheet'
+import OsblFinance from '@/components/pages/contribution/new/OsblFinances'
 import OsblDocuments from '@/components/pages/contribution/new/OsblDocuments'
 import OsblLocations from '@/components/pages/contribution/new/OsblLocations'
 import { CurrentUserType } from '@/types/types'
