+++ app/frontend/tests/forms/SignInForm.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach, Mock } from 'vitest'
 import { render, screen, fireEvent } from '@testing-library/react'
-import SignInForm from '@/components/forms/SignInForm'
+import SignInForm from '@/components/pages/auth/signIn/SignInForm'
 import useFormHandler from '@/hooks/useFormHandler'
 
 // Mock sessionStorage
